Finished save Calculation. Need to add confirmation.

+++ app/src/main/java/az/squareroot/customstaxescalc/ItemClickListener.kt
@@ -1,5 +1,5 @@
 package az.squareroot.customstaxescalc
 
 interface ItemClickListener {
-    fun onClick(id: Long)
+    fun onClick(id: Int)
 }
